generator/pyproject.toml
[project]
name = "soundflow-generator"
version = "0.1.0"
description = "Daily track generation + upload + catalog writer for SoundFlow."
requires-python = ">=3.10"
dependencies = [
  "boto3>=1.34.0",
  "botocore>=1.34.0",
  "orjson>=3.10.0",
  "pyyaml>=6.0.1",
  "pydub>=0.25.1",
  "numpy>=1.26.0",
  # Optional (premium GPU):
  "torch>=2.2.0",
  "transformers>=4.44.0",
  "accelerate>=0.33.0",
]

[tool.uv]
# Generate lockfile with: uv lock
# Install deps with: uv sync

[tool.ruff]
line-length = 100
target-version = "py310"
